requirements.txt
# JobMatch AI dependencies (Python 3.12)
# Versions are pinned for reproducible setup across different machines.

# API server
fastapi==0.115.12
uvicorn[standard]==0.34.2

# HTTP client for Cloudflare Workers AI
httpx==0.28.1

# CV parsing
PyPDF2==3.0.1
python-docx==1.1.2
python-multipart==0.0.20

# Configuration and validation
pydantic==2.11.3
pydantic-settings==2.8.1
python-dotenv==1.1.0
